% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchMlr3Shiny.R
\name{launchMlr3Shiny}
\alias{launchMlr3Shiny}
\title{Title Launch mlr3shiny}
\usage{
launchMlr3Shiny(test = FALSE)
}
\arguments{
\item{test}{Argument used for testing purposes. Default is FALSE.}
}
\value{
None
}
\description{
Launch an instance of mlr3shiny to perform machine learning in Shiny using mlr3.
}
\examples{
<<<<<<< HEAD
if(interactive()){
launchMlr3Shiny()
=======
if (interactive()) {
   launchMlr3Shiny()
>>>>>>> b3a20c2f93a445bca1de68530952c3fd869a4beb
}

}
