% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchMlr3Shiny.R
\name{launchMlr3Shiny}
\alias{launchMlr3Shiny}
\title{Title Launch mlr3shiny}
\usage{
launchMlr3Shiny(test = FALSE)
}
\arguments{
\item{test}{Argument used for testing purposes. Default is FALSE.}
}
\description{
Launch an instance of mlr3shiny to perform machine learning in Shiny using mlr3.
}
\examples{
\dontrun{
launchMlr3Shiny()
}

}
